se viene eliminato un bookable_service vengono eliminate anche le reservation che si riferivano al bookable_service eliminato

+++ routes/bookable_services.js
@@ -143,6 +143,7 @@ router.delete('/:id/', async (req, res) => {
 	
 	try {
 		await Bookable_service.deleteOne({ _id: req.params.id });
+		await Reservation.deleteMany({ bookable_service: req.params.id });
 		res.status(200).end();
 	}
 	catch (err) {
@@ -282,6 +283,17 @@ router.get('/reservation', async (req, res) => {
 router.delete('/reservation/:id/', async (req, res) => {
 	
 	try {
+		const reservation = await Reservation.find({ _id: req.params.id });
+
+		if (reservation.length === 0) {
+			throw new Error(`The reservation with id ${req.params.id} doesn't exitst`);
+		}
+
+		const bs = await Bookable_service.find({ '_id': reservation[0].bookable_service });
+
+		bs[0].reservation_left += reservation[0].qty;
+		await bs[0].save();
+
 		await Reservation.deleteOne({ _id: req.params.id });
 		res.status(200).end();
 	}
